Share transfer helpers
